requirements.txt
flask
flask-sqlalchemy
flask-login
flask-wtf
flask-migrate
email-validator
python-dotenv
python-slugify
openai
selenium
